Refine waveform slide module boundaries
